+++ google/cloud/bigtable/google_cloud_cpp_bigtable.bzl
@@ -39,7 +39,6 @@ google_cloud_cpp_bigtable_hdrs = [
     "admin/internal/bigtable_table_admin_retry_traits.h",
     "admin/internal/bigtable_table_admin_stub.h",
     "admin/internal/bigtable_table_admin_stub_factory.h",
-    "admin/internal/wait_for_consistency.h",
     "admin_client.h",
     "app_profile_config.h",
     "async_row_reader.h",
@@ -80,6 +79,7 @@ google_cloud_cpp_bigtable_hdrs = [
     "internal/rpc_policy_parameters.h",
     "internal/rpc_policy_parameters.inc",
     "internal/unary_client_utils.h",
+    "internal/wait_for_consistency.h",
     "metadata_update_policy.h",
     "mutation_batcher.h",
     "mutations.h",
@@ -121,7 +121,6 @@ google_cloud_cpp_bigtable_srcs = [
     "admin/internal/bigtable_table_admin_option_defaults.cc",
     "admin/internal/bigtable_table_admin_stub.cc",
     "admin/internal/bigtable_table_admin_stub_factory.cc",
-    "admin/internal/wait_for_consistency.cc",
     "admin_client.cc",
     "app_profile_config.cc",
     "client_options.cc",
@@ -147,6 +146,7 @@ google_cloud_cpp_bigtable_srcs = [
     "internal/prefix_range_end.cc",
     "internal/readrowsparser.cc",
     "internal/rowreaderiterator.cc",
+    "internal/wait_for_consistency.cc",
     "metadata_update_policy.cc",
     "mutation_batcher.cc",
     "mutations.cc",
